Student_Database_Management_System/requirements.txt
pyodbc
tkcalendar
python-dotenv